requirements.txt
python-docx
pillow
pytesseract
openai
pandas
fastapi
uvicorn
python-multipart
python-dateutil
groq
